remove hacky_wrapper from BackendSelect (#49079)
Summary:
Pull Request resolved: https://github.com/pytorch/pytorch/pull/49079

BackendSelect kernels have been changed to be written the new way, so this hacky_wrapper here isn't needed anymore.
This PR is not expected to change perf or anything, just simplify the code a bit. The hacky_wrapper here was a no-op and not creating any actual wrappers
because it short-cirtuits to not create a wrapper when there is no wrapper needed.
ghstack-source-id: 118318436

Test Plan: waitforsandcastle

Reviewed By: bhosmer

Differential Revision: D25421633

fbshipit-source-id: 7a6125613f465dabed155dd892c8be6af5c617cf