Fix shape inference for multiple outputs with different output dtypes (#52417)
Summary:
Pull Request resolved: https://github.com/pytorch/pytorch/pull/52417

When we have multiple outputs, previously, we will set `infered_data_type` to the first output and stick to it. This is not correct if we have more outputs that has different dtype. The fix here will revert `infered_data_type` back to previous value (`UNDEFINED`) so that we can still enter the conditional check and set the right dtype for second and more outputs.

Test Plan:
```
buck test caffe2/caffe2/fb/operators:infer_bound_shape_op_test
```

Reviewed By: khabinov

Differential Revision: D26502161

fbshipit-source-id: 647f0106a5785dc156fddfc196ac67001602fda8