[package] Add `deny` method to `PackageExporter` (#53233)
Summary:
Pull Request resolved: https://github.com/pytorch/pytorch/pull/53233

**Summary**
This commit adds a `deny` method to `PackageExporter` that allows
modules to be prohibited during the packaging process. A dependency on a
module matching the names or globs that `deny` was called with will
cause an exception to be raised.

**Test Plan**
This commit adds unit tests to `PackagingTest` for this new method:
`test_deny` and `test_deny_glob`.

**Fixes**
This commit fixes #53217.

Test Plan: Imported from OSS

Reviewed By: suo

Differential Revision: D26834010

Pulled By: SplitInfinity

fbshipit-source-id: 469b5c6741bcc6dab77e352f41db38fa1e0dae12